Omit runninghead and footer elements in the text
Elements with the semantic action runninghead/footer are now showed only
on the top/bottom lines. They used to appear unformatted in the text,
but that was not very useful. If you want a header to appear in the text,
use the 'runningHead' option in a style. Otherwise you need two elements.