Trying to fix the test suite for VPATH builds
According to https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=821937
the test suite fails on the Debian build server. The reason is that it
cannot cd into the test directory. This fix should make the test suite a
bit more robust for VPATH build scenarios and hopefully fix the bug