Use the input value instead of the private fn _lou_getLastTableList
`_lou_getLastTableList` is part of the private API of liblouis which
is no longer exported. The function seems to just grab the name of the
last compiled table, which in this case should be the input of the
last call to `lou_getTable`. This input is `nested->value`. So I'm
just using that instead.