Add sanitizer checks to CI
this is largely inspired by #78. It just moves the jobs into a
separate workflow and uses a build matrix to run the different
sanitizers, similarly to how it is done in liblouis.

One draw-back is that we define the LIBLOUIS_VERSION in both
workflows. OTOH it seems nicer to keep the main build and the
sanitizer as separate workflows.