Fix boldface indicator with English and Greek letters
If an English lowercase or capital letter, or a Greek lowercase, math
variant or capital letter appears within a <mi
mathvariant="bold">letter</mi> element, now it is preceded by a Nemeth
boldface indicator. Moreover, the contracted form of a horizontal bar
over or under a single letter is extended to Greek letters and
boldface English and Greek letters.

Co-authored-by: Norbert Márkus <hamilfonz@gmail.com>

Signed-off-by: Attila Hammer <hammera@pickup.hu>